requirements.txt
pytesseract==0.3.5
Pillow==7.2.0
pyperclip==1.8.0